First attempt Calendar refacto
